+++ gulpfile.js
@@ -2,4 +2,4 @@ const gulp = require('gulp');
 const { build } = require('@justeat/gulp-build-fozzie');
 
 // Grab the Fozzie build tasks
-build(gulp, {});
+build(gulp);