+++ src/js/index.js
@@ -1,7 +1,7 @@
 // All helper functions will be imported here, so that they can all be exported within one object.
-import { getBreakpoints, currentScreenWidth } from './modules/breakpointHelper';
+import { getBreakpoints, getCurrentScreenWidth } from './modules/breakpointHelper';
 
 export {
     getBreakpoints,
-    currentScreenWidth
+    getCurrentScreenWidth
 };